The following changes have been made:
Improve “Assign Sequence Number” function. if an attribute does not have a "sequenceNumber" tagged value, this function will create a new “sequenceNumber” tagged value and assign a unique number to it. 

Improve “Run Conformance Test” function. User can locate the error element in the project browser by double clicking on the data cell in the report window. The function will also check whether an element ( Class, Package, Connector, Attribute ) has any notes associate with it or not. If the element does not have any notes, a warning message will be displayed in the report window. 

Check source control status. If the model is source controlled, all the editing functions will be available only when the source control status is “Checked out”.


git-svn-id: https://svnserv.csiro.au/svn/ext/HollowWorld@1893 96135fb9-dd2c-0410-bee2-84220cfab6eb